Add ConsoleApp project with .NET 8.0 support
Updated the solution file to include a new "ConsoleApp" project with Debug and Release configurations. The project file is set up to use the .NET SDK, targeting .NET 8.0, with implicit usings and nullable reference types enabled. Added a package reference to `Microsoft.Extensions.Hosting` version 9.0.3. The `Program.cs` file was modified to create and run the application host.